Refactoring
- Modified the way data is read/written from files
    - When reading you now specify a position to read from
    - When writing you no longer specify offset and length
    - This simplified a lot of the code that depends on reading/writing files
- Added test coverage for writing/reading entries to/from file
    - This tests the data extension methods and verifies custom serlaiization
- Fixed specification tests to clean up files before running
    - This helps clean up any files left over as residuals from failed tests

+++ Streaks.Spec/Basics/Basics.steps.cs
@@ -21,6 +21,8 @@ namespace Streaks.Test.Specifications
         [BeforeScenario]
         public void Setup()
         {
+            if (Directory.Exists(Path)) Directory.Delete(Path, true);
+
             streak = Streak.Open(Path);
 
             writer = streak.Writer();
@@ -35,7 +37,7 @@ namespace Streaks.Test.Specifications
             reader.Dispose();
             writer.Dispose();
 
-            Directory.Delete(Path, true);
+            if (Directory.Exists(Path)) Directory.Delete(Path, true);
         }
 
         [Given(@"an empty streak")]
